Modify text rendering to have multiple passes and cleanup
So far, text rendering in the capture window was done on top of all
other rendering. With this, the overlay that was rendered for the
iterator UI, would not affect the text on the timer slices, which made
it harder to discern for users which text belonged where. After the
changes, text is rendered in multiple passes to ensure that overlays
affect the color of the text below. To make this a reasonably clean
change, some changes to rendering of various UI elements and their
z values used were needed.

Fix order of rendering and z values used.